feat: Implement segment loading and memory management improvements
- Added `load_segments` method in `ElfLoader` to load ELF segments into memory, including BSS initialization and page flag application.
- Refactored memory allocation logic in `LockedBuddyHeap` to improve clarity.
- Introduced `split_and_reinsert_regions` method in `ProcessAddressSpace` for better memory protection handling in `mprotect`.
- Enhanced NUMA hint scanning with `scan_region_numa_hints` for improved memory locality.
- Updated `check_anon_constraints` to use a bounded sender for swap entry checks in `async_swapout`.
- Modified fast allocator methods to use `PerCpuFastMagazine` for better performance.
- Reconfigured NUMA allocator to use `NumaPmmAllocator` for improved memory management.
- Implemented aging logic in `MglruList` for better page reclamation strategies.
- Added `try_process_dns_answer` method in `MdnsService` for streamlined DNS response handling.
- Refactored command evaluation in `ExoShell` for improved command execution flow.
- Introduced global LRU page finding in `PageCache` to optimize page eviction logic.
- Added graphics framebuffer tests to the kernel suite for enhanced graphics functionality validation.

+++ kernel/src/graphics/framebuffer/qemu_tests.rs
@@ -1325,7 +1325,7 @@ pub fn wave9_fill_circle_no_gaps_smoke() -> bool {
     let mut fb = unsafe { Framebuffer::new(info) };
     fb.enable_double_buffering_from_vec(vec![0u32; (w * h) as usize]);
     let (cx, cy, r) = (16i32, 16i32, 7i32);
-    fb.fill_circle(cx, cy, r as u32, Color::with_alpha(0, 0, 255, 255));
+    fb.fill_circle(cx, cy, r, Color::with_alpha(0, 0, 255, 255));
     let buf = fb.back_buffer.as_ref().unwrap();
     for y in (cy - r)..=(cy + r) {
         for x in (cx - r)..=(cx + r) {
